env: mv vmctl-init-conf into a env command
What was in vmctl-init-conf was not only for creating the configuration
directory but for entering a vmctl environment without installing it.
Move everything to cmd/env and add the env command to vmclt

To bootstrap a configuration dir run `vmctl env --confdir DIR --bootstrap`
this will not execute an environment. `vmctl env --confdir DIR` will
enter a new vmctl environment with the 'exec' command

Update the README to reflect the changes in vmctl initial setup

Signed-off-by: Joel Granados <joel.granados@kernel.org>